@@ -26,21 +26,20 @@
     <properties>
         <jar.name>spring-boot-migrator</jar.name>
         <jaxb-maven-plugin.version>2.5.0</jaxb-maven-plugin.version>
-        <maven.compiler.target>11</maven.compiler.target>
-        <maven.compiler.source>11</maven.compiler.source>
-        <openrewrite.version>7.22.0</openrewrite.version>
-        <openrewrite.spring.version>4.17.0</openrewrite.spring.version>
-        <spring-boot.version>2.5.12</spring-boot.version>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <java.version>17</java.version>
+        <spring-shell.version>2.1.0-RC1</spring-shell.version>
+        <openrewrite.version>7.26.0</openrewrite.version>
+        <openrewrite.spring.version>4.23.0</openrewrite.spring.version>
+        <spring-boot.version>2.7.1</spring-boot.version>
         <progressbar.version>0.9.0</progressbar.version>
         <testcontainers.version>9baedef17589ff70f7a585c809e0b1beb27eff62</testcontainers.version>
-        <maven-invoker.version>3.0.1</maven-invoker.version>
-        <spring-shell-starter.version>2.0.1.RELEASE</spring-shell-starter.version>
+        <maven-invoker.version>3.2.0</maven-invoker.version>
         <shrinkwrap.resolvers.version>3.1.4</shrinkwrap.resolvers.version>
-        <lombok.version>1.18.20</lombok.version>
+        <lombok.version>1.18.24</lombok.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.csvParser.outputEncoding>UTF-8</project.csvParser.outputEncoding>
-        <jackson.version>2.12.6</jackson.version>
-        <jackson-databind.version>2.12.6.1</jackson-databind.version>
         <generated-sources.dir>src/generated/java</generated-sources.dir>
     </properties>
 
@@ -370,8 +382,12 @@
                 </configuration>
             </plugin>
             <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
                 <configuration>
+                    <source>17</source>
+                    <target>17</target>
                     <forceJavacCompilerUse>true</forceJavacCompilerUse>
                 </configuration>
             </plugin>
