@@ -82,13 +82,6 @@ public class BootifySimpleMuleAppIntegrationTest extends IntegrationTestBaseClas
 
     @Test
     @Tag("integration")
-    @DisabledIfSystemProperty(named= "os.arch", matches = "aarch64", disabledReason = "imbcom/mq image not supported with Apple Silicon")
-    void t1_testWebsphereMqMigration() throws JMSException, InterruptedException {
-        checkWmqIntegration(rabbitMqContainer.getNetwork());
-    }
-
-    @Test
-    @Tag("integration")
     public void  t0_springIntegrationWorks() throws IOException, TimeoutException, InterruptedException {
         intializeTestProject();
         scanProject();
@@ -109,10 +102,15 @@ public class BootifySimpleMuleAppIntegrationTest extends IntegrationTestBaseClas
         checkSendHttpMessage(container.getContainer().getMappedPort(9081));
         checkInboundGatewayHttpMessage(container.getContainer().getMappedPort(9081));
         checkRabbitMqIntegration(ampqChannel);
+        checkDbIntegration(container.getContainer().getMappedPort(9081));
     }
 
-
-
+    @Test
+    @Tag("integration")
+    @DisabledIfSystemProperty(named= "os.arch", matches = "aarch64", disabledReason = "imbcom/mq image not supported with Apple Silicon")
+    void t1_testWebsphereMqMigration() throws JMSException, InterruptedException {
+        checkWmqIntegration(rabbitMqContainer.getNetwork());
+    }
 
     private void checkRabbitMqIntegration(Channel amqpChannel)
             throws IOException, InterruptedException {
@@ -155,6 +153,12 @@ public class BootifySimpleMuleAppIntegrationTest extends IntegrationTestBaseClas
         jmsSender.sendMessage(mappedPort, "DEV.QUEUE.1", "Test WMQ message");
         boolean latchResult = latch.await(1000000, TimeUnit.MILLISECONDS);
         assertThat(latchResult).isTrue();
+    }
+
+    private void checkDbIntegration(int port) {
+        ResponseEntity<String> responseEntity = restTemplate.getForEntity("http://localhost:" + port + "/db", String.class);
+        assertThat(responseEntity.getStatusCode()).isEqualTo(HttpStatus.OK);
+        assertThat(responseEntity.getBody()).contains("{\"ID\":1,\"USERNAME\":\"TestUser\",\"PASSWORD\":\"secret\"");
     }
 
     private void checkInboundGatewayHttpMessage(int port) {
