@@ -17,20 +17,22 @@
 package org.springframework.sbm.common.filter;
 
 import lombok.RequiredArgsConstructor;
+import org.openrewrite.SourceFile;
 import org.springframework.sbm.project.resource.ProjectResource;
 import org.springframework.sbm.project.resource.ProjectResourceSet;
+import org.springframework.sbm.project.resource.RewriteSourceFileHolder;
 import org.springframework.sbm.project.resource.filter.ProjectResourceFinder;
 
 import java.nio.file.Path;
 import java.util.Optional;
 
 @RequiredArgsConstructor
-public class AbsolutePathResourceFinder implements ProjectResourceFinder<Optional<ProjectResource>> {
+public class AbsolutePathResourceFinder implements ProjectResourceFinder<Optional<RewriteSourceFileHolder<? extends SourceFile>>> {
 
     private final Path absoluteResourcePath;
 
     @Override
-    public Optional<ProjectResource> apply(ProjectResourceSet projectResourceSet) {
+    public Optional<RewriteSourceFileHolder<? extends SourceFile>> apply(ProjectResourceSet projectResourceSet) {
         if (absoluteResourcePath == null || ! absoluteResourcePath.isAbsolute()) {
             throw new IllegalArgumentException("Given path '"+absoluteResourcePath+"' is not absolute");
         }
@@ -38,7 +40,6 @@ public class AbsolutePathResourceFinder implements ProjectResourceFinder<Optiona
         return projectResourceSet
                 .stream()
                 .filter(r -> searchForPath.equals(r.getAbsolutePath()))
-                .map(ProjectResource.class::cast)
                 .findFirst();
 
     }
