@@ -16,10 +16,13 @@
 package org.springframework.sbm.java.impl;
 
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.exception.ExceptionContext;
+import org.openrewrite.ExecutionContext;
 import org.openrewrite.Recipe;
 import org.openrewrite.java.ChangeType;
 import org.openrewrite.java.MethodMatcher;
 import org.openrewrite.java.search.FindMethods;
+import org.openrewrite.java.search.UsesType;
 import org.openrewrite.java.tree.J;
 import org.openrewrite.java.tree.JavaType;
 import org.openrewrite.java.tree.TypeUtils;
@@ -28,10 +31,12 @@ import org.springframework.sbm.java.filter.JavaSourceListFilter;
 import org.springframework.sbm.java.refactoring.JavaGlobalRefactoring;
 import org.springframework.sbm.project.resource.ProjectResourceSet;
 import org.springframework.sbm.project.resource.RewriteSourceFileHolder;
+import org.springframework.sbm.support.openrewrite.GenericOpenRewriteRecipe;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Slf4j
@@ -95,7 +100,7 @@ public class ProjectJavaSourcesImpl implements ProjectJavaSources {
     @Override
     public List<MethodCall> findMethodCalls(String methodPattern) {
         List<MethodCall> matches = new ArrayList<>();
-        FindMethods findMethods = new FindMethods(methodPattern, true);
+        FindMethods findMethods = new FindMethods(methodPattern,true, null);
         MethodMatcher methodMatcher = new MethodMatcher(methodPattern);
         find(findMethods).stream()
                 .map(m -> list().stream().filter(js -> js.getResource().getId().equals(m.getId())).findFirst().get())
@@ -119,6 +124,18 @@ public class ProjectJavaSourcesImpl implements ProjectJavaSources {
         return matches;
     }
 
+    @Override
+    public List<? extends JavaSource> findClassesUsingType(String fqName) {
+        UsesType<ExecutionContext> usesType = new UsesType<>(fqName);
+        GenericOpenRewriteRecipe<UsesType<ExecutionContext>> recipe = new GenericOpenRewriteRecipe<>(() -> usesType);
+        return find(recipe).stream()
+                .filter(RewriteSourceFileHolder.class::isInstance)
+                .map(RewriteSourceFileHolder.class::cast)
+                .filter(r -> r.getType().isAssignableFrom(J.CompilationUnit.class))
+                .map(r -> (OpenRewriteJavaSource)r)
+                .collect(Collectors.toList());
+    }
+
     private boolean hasTypeImplementing(J.ClassDeclaration c, String type) {
         try {
             return c.getImplements() != null &&
@@ -128,8 +145,8 @@ public class ProjectJavaSourcesImpl implements ProjectJavaSources {
                                         JavaType.FullyQualified fullyQualified = TypeUtils.asFullyQualified(
                                                 intaface.getType()
                                         );
-                                        if(fullyQualified == null) {
-                                            log.error("Could not resolve implemented type '" + ((J.Identifier)intaface).getSimpleName() + "'");
+                                        if(fullyQualified == null && J.Identifier.class.isInstance(intaface)) {
+                                            log.error(String.format("Could not calculate if class '%s' implements an interface compatible to '%s'. Type of interface '%s' could not be resolved and was '%s'", c, type, intaface, intaface.getType().toString()));
                                             return false;
                                         }
                                         return fullyQualified
