@@ -62,14 +62,13 @@ public class ResponseBuilderTest {
         String expected = ""
                 + "import java.util.Set;\n"
                 + "\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
-                + "\n"
                 + "import org.springframework.http.HttpMethod;\n"
+                + "import org.springframework.http.ResponseEntity;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        b.allow(HttpMethod.resolve(\"POST\"), HttpMethod.resolve(\"PUT\"));\n"
                 + "        b.allow(Set.of(\"GET\").stream().map(HttpMethod::resolve).toArray(String[]::new));\n"
                 + "        return b;\n"
@@ -111,13 +110,14 @@ public class ResponseBuilderTest {
                 + "";
 
         String expected = ""
+                + "import org.springframework.http.ResponseEntity;\n"
+                + "\n"
                 + "import java.util.Date;\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        b.headers(h -> h.setExpires(new Date(100000).toInstant()));\n"
                 + "        return b;\n"
                 + "    }\n"
@@ -158,14 +158,13 @@ public class ResponseBuilderTest {
         String expected = ""
                 + "import java.util.Locale;\n"
                 + "\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
-                + "\n"
                 + "import org.springframework.http.HttpHeaders;\n"
+                + "import org.springframework.http.ResponseEntity;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        b.headers(h -> h.set(HttpHeaders.CONTENT_LANGUAGE, \"ua\"));\n"
                 + "        b.headers(h -> h.setContentLanguage(Locale.ITALY));\n"
                 + "        return b;\n"
@@ -174,7 +173,9 @@ public class ResponseBuilderTest {
                 + "";
 
         ProjectContext projectContext = TestProjectContext.buildProjectContext()
-                .withBuildFileHavingDependencies("javax:javaee-api:8.0")
+                .withBuildFileHavingDependencies(
+                        "javax:javaee-api:8.0",
+                        "org.springframework:spring-web:5.3.18")
                 .withJavaSources(javaSource)
                 .build();
 
@@ -204,13 +205,14 @@ public class ResponseBuilderTest {
                 + "";
 
         String expected = ""
+                + "import org.springframework.http.ResponseEntity;\n"
+                + "\n"
                 + "import java.util.Date;\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        b.lastModified(new Date(100000).toInstant());\n"
                 + "        return b;\n"
                 + "    }\n"
@@ -249,13 +251,14 @@ public class ResponseBuilderTest {
                 + "";
 
         String expected = ""
+                + "import org.springframework.http.ResponseEntity;\n"
+                + "\n"
                 + "import javax.ws.rs.core.MultivaluedMap;\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        MultivaluedMap m;\n"
                 + "        b.headers(h -> {\n"
                 + "            h.clear();\n"
@@ -298,12 +301,12 @@ public class ResponseBuilderTest {
                 + "";
 
         String expected = ""
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
+                + "import org.springframework.http.ResponseEntity;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "       BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "       ResponseEntity.BodyBuilder b;\n"
                 + "       b.eTag(\"foo\");\n"
                 + "       return b;\n"
                 + "    }\n"
@@ -346,13 +349,14 @@ public class ResponseBuilderTest {
                 + "";
 
         String expected = ""
+                + "import org.springframework.http.ResponseEntity;\n"
+                + "\n"
                 + "import javax.ws.rs.core.MediaType;\n"
-                + "import org.springframework.http.ResponseEntity.BodyBuilder;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
-                + "    public BodyBuilder test() {\n"
-                + "        BodyBuilder b;\n"
+                + "    public ResponseEntity.BodyBuilder test() {\n"
+                + "        ResponseEntity.BodyBuilder b;\n"
                 + "        b.contentType(MediaType.APPLICATION_JSON_TYPE);\n"
                 + "        b.headers(h -> h.set(HttpHeaders.CONTENT_TYPE, \"application/json\"));\n"
                 + "        return b;\n"
