@@ -59,20 +59,19 @@ public class ResponseStatusFamilyTest {
 
         String expected = ""
                 + "import org.springframework.http.HttpStatus;\n"
-                + "\n"
                 + "import org.springframework.http.HttpStatus.Series;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
                 + "    public void test() {\n"
-                + "       Series f1 = Series.INFORMATIONAL;\n"
-                + "       Series f2 = Series.SUCCESSFUL;\n"
-                + "       Series f3 = Series.REDIRECTION;\n"
-                + "       Series f4 = Series.CLIENT_ERROR;\n"
-                + "       Series f5 = Series.SERVER_ERROR;\n"
+                + "       HttpStatus.Series f1 = HttpStatus.Series.INFORMATIONAL;\n"
+                + "       HttpStatus.Series f2 = HttpStatus.Series.SUCCESSFUL;\n"
+                + "       HttpStatus.Series f3 = HttpStatus.Series.REDIRECTION;\n"
+                + "       HttpStatus.Series f4 = HttpStatus.Series.CLIENT_ERROR;\n"
+                + "       HttpStatus.Series f5 = HttpStatus.Series.SERVER_ERROR;\n"
                 + "       \n"
                 + "       int code = 201;\n"
-                + "       Series custom = HttpStatus.Series.resolve(code);\n" // FIXME: #116
+                + "       HttpStatus.Series custom = HttpStatus.Series.resolve(code);\n"
                 + "    }\n"
                 + "}\n"
                 + "";
@@ -108,12 +107,13 @@ public class ResponseStatusFamilyTest {
                 + "";
 
         String expected = ""
+                + "import org.springframework.http.HttpStatus;\n"
                 + "import org.springframework.http.HttpStatus.Series;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
                 + "    public void test() {\n"
-                + "       Series f = Series.OTHER;\n"
+                + "       HttpStatus.Series f = HttpStatus.Series.OTHER;\n"
                 + "    }\n"
                 + "}\n"
                 + "";
@@ -148,15 +148,13 @@ public class ResponseStatusFamilyTest {
                 + "";
 
         String expected = ""
-                + "import org.springframework.http.HttpStatus.Series;\n"
-                + "\n"
                 + "import org.springframework.http.HttpStatus;\n"
                 + "\n"
                 + "public class TestController {\n"
                 + "\n"
                 + "    public void test() {\n"
                 + "       int code = 201;\n"
-                + "       Series custom = HttpStatus.Series.resolve(code);\n"  // FIXME: #116
+                + "       HttpStatus.Series custom = HttpStatus.Series.resolve(code);\n"
                 + "    }\n"
                 + "}\n"
                 + "";
