@@ -40,8 +40,10 @@ class SelectTranslatorTest {
         DslSnippet output = target.translate(1, input, null, null, null, null);
 
         assertThat(output.getRenderedSnippet()).isEqualTo(
-                        " // TODO: substitute expression language with appropriate java code \n" +
-                                " // TODO: use appropriate translation for pagination for more information visit: https://bit.ly/3xlqByv \n" +
-                                "                .handle((p, h) -> jdbcTemplate.queryForList(\"Select * from Students where name like \\\"Sandeep\\\"\"))");
+                        "// TODO: substitute expression language with appropriate java code \n" +
+                                "// TODO: The datatype might not be LinkedMultiValueMap please substitute the right type for payload\n" +
+                                ".<LinkedMultiValueMap<String, String>>handle((p, h) ->\n" +
+                                "                        jdbcTemplate.queryForList(\n" +
+                                "                                \"Select * from Students where name like \\\"Sandeep\\\"\"))");
     }
 }
